DAY 10/backend_dynamic_role_recommender/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai
langchain
langgraph
faiss-cpu
pdfminer.six
gitpython
beautifulsoup4
httpx
tavily-python
python-dotenv
langchain-community


